moved some stuff around to better reflect the code organization, hopefully
