+++ CompatBot/Commands/Sudo.Mod.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 using System.Threading.Tasks;
-using CompatBot.Providers;
+using CompatBot.Database.Providers;
 using CompatBot.Utils;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
