+++ CompatBot/EventHandlers/LogInfoHandler.cs
@@ -4,8 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using CompatApiClient;
 using CompatBot.Commands;
-using CompatBot.LogParsing;
-using CompatBot.LogParsing.SourceHandlers;
+using CompatBot.EventHandlers.LogParsing;
+using CompatBot.EventHandlers.LogParsing.POCOs;
+using CompatBot.EventHandlers.LogParsing.SourceHandlers;
 using CompatBot.ResultFormatters;
 using CompatBot.Utils;
 using DSharpPlus;
