+++ CompatBot/EventHandlers/LogParsing/SourceHandlers/PlainText.cs
@@ -4,7 +4,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using DSharpPlus.Entities;
 
-namespace CompatBot.LogParsing.SourceHandlers
+namespace CompatBot.EventHandlers.LogParsing.SourceHandlers
 {
     public class PlainTextHandler: ISourceHandler
     {
@@ -26,7 +26,7 @@ namespace CompatBot.LogParsing.SourceHandlers
                     {
                         var memory = writer.GetMemory(Config.MinimumBufferSize);
                         read = await stream.ReadAsync(memory, Config.Cts.Token);
-                        writer.Advance(read); //todo: test that .Advance(0) works as expected
+                        writer.Advance(read);
                         flushed = await writer.FlushAsync(Config.Cts.Token).ConfigureAwait(false);
                     } while (read > 0 && !(flushed.IsCompleted || flushed.IsCanceled || Config.Cts.IsCancellationRequested));
                 }
