+++ CompatBot/EventHandlers/LogParsing/SourceHandlers/ZipHandler.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 using CompatBot.Utils;
 using DSharpPlus.Entities;
 
-namespace CompatBot.LogParsing.SourceHandlers
+namespace CompatBot.EventHandlers.LogParsing.SourceHandlers
 {
     public class ZipHandler: ISourceHandler
     {
@@ -65,7 +65,7 @@ namespace CompatBot.LogParsing.SourceHandlers
                             {
                                 var memory = writer.GetMemory(Config.MinimumBufferSize);
                                 read = await zipStream.ReadAsync(memory, Config.Cts.Token);
-                                writer.Advance(read); //todo: test that .Advance(0) works as expected
+                                writer.Advance(read);
                                 flushed = await writer.FlushAsync(Config.Cts.Token).ConfigureAwait(false);
                             } while (read > 0 && !(flushed.IsCompleted || flushed.IsCanceled || Config.Cts.IsCancellationRequested));
                         }
